+++ packages/core/src/grammar/index.ts
@@ -1,0 +1,4 @@
+export * from './inflect'
+
+export * from './constants'
+export * from './types'