+++ packages/core/src/inflection/index.ts
@@ -1,5 +1,0 @@
-export * from './decline'
-export * from './declineForNumber'
-
-export * from './constants'
-export * from './types'