+++ packages/core/src/localization/index.ts
@@ -1,5 +1,4 @@
 export { default as localizeMessage } from './localizeMessage'
 export { default as localizeUrl } from './localizeUrl'
-export { default as prepareFormat } from './prepareFormat'
 
 export * from './types'